requirements.txt
Flask>=3.0,<4
Werkzeug>=3.0,<4
Flask-SQLAlchemy>=3.1,<3.2
requests>=2.31,<3
python-dotenv>=1.0,<2
tzdata>=2024.1; sys_platform == "win32"
backports.zoneinfo; python_version < "3.9"
